Add an object dictionnary
