+++ trampoline_unity/trampoline/Assets/Scripts/Dictionnary.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2df58a621b6aab459560390969d0793
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
